--- vllm/model_executor/parallel_utils/README.md
@@ -1,1 +1,0 @@
-The files in this folder are ported from [Megatron-LM](https://github.com/NVIDIA/Megatron-LM/tree/main/megatron/core). We only keep the codes that are used in inference.
